requirements.txt
PyYAML==6.0.2
pypdf==3.14.0
requests==2.32.3
pycryptodome==3.15.0
iiif-prezi3==1.2.1
PyMuPDF==1.24.11
openai-whisper==20240930
mutagen==1.47.0
ffmpeg-python==0.2.0
pyvips==2.2.3
pandas==2.2.3
pytest
